+++ src/main/java/com/br/kerberus/urd/controller/ServerController.java
@@ -2,7 +2,7 @@ package com.br.kerberus.urd.controller;
 
 import com.br.kerberus.urd.exception.ManagedException;
 import com.br.kerberus.urd.exception.NoManagedException;
-import com.br.kerberus.urd.entity.LogHttpMessages;
+import com.br.kerberus.urd.entity.core.LogHttpMessages;
 import com.br.kerberus.urd.resource.ErrorResponse;
 import com.br.kerberus.urd.resource.ServerRequest;
 import com.br.kerberus.urd.resource.ServerResponse;
@@ -16,7 +16,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Api(value = "server")
@@ -39,8 +38,7 @@ public class ServerController {
     @LogHttpMessages
     @ResponseStatus(value = HttpStatus.CREATED)
     @PostMapping(path = {"/add"})
-    public ResponseEntity<ServerResponse> addServer(HttpServletRequest request,
-                                                    @RequestBody @Validated ServerRequest server) throws NoManagedException {
+    public ResponseEntity<ServerResponse> addServer(@RequestBody @Validated ServerRequest server) throws NoManagedException {
 
         ServerResponse response = new ServerResponse(service.addServer(server));
 
@@ -57,8 +55,7 @@ public class ServerController {
     @LogHttpMessages
     @ResponseStatus(value = HttpStatus.OK)
     @PutMapping(path = {"/{id}"})
-    public ResponseEntity<ServerResponse> updateServer(HttpServletRequest request,
-                                                       @RequestBody @Validated ServerResponse server,
+    public ResponseEntity<ServerResponse> updateServer(@RequestBody @Validated ServerResponse server,
                                                        @PathVariable(name = "id", required = true) Long id) {
 
         return ResponseEntity.status(HttpStatus.OK).body(null);
@@ -74,8 +71,7 @@ public class ServerController {
     @LogHttpMessages
     @ResponseStatus(value = HttpStatus.OK)
     @DeleteMapping(path = {"/{id}"})
-    public ResponseEntity<ServerResponse> deleteServer(HttpServletRequest request,
-                                                       @PathVariable(name = "id", required = true) Long id) {
+    public ResponseEntity<ServerResponse> deleteServer(@PathVariable(name = "id", required = true) Long id) {
 
         return ResponseEntity.status(HttpStatus.OK).body(null);
     }
@@ -105,8 +101,7 @@ public class ServerController {
     @LogHttpMessages
     @ResponseStatus(value = HttpStatus.OK)
     @GetMapping(path = {"/{id}"})
-    public ResponseEntity<ServerResponse> getServerById(HttpServletRequest request,
-                                                        @PathVariable(name = "id") Integer id) throws ManagedException {
+    public ResponseEntity<ServerResponse> getServerById(@PathVariable(name = "id") Integer id) throws ManagedException {
 
         ServerResponse server = new ServerResponse(service.getServerById(id));
 
@@ -123,8 +118,7 @@ public class ServerController {
     @LogHttpMessages
     @ResponseStatus(value = HttpStatus.OK)
     @GetMapping(path = {"/search/byHostname/{name}"})
-    public ResponseEntity<ServerResponse> getServerByHostname(HttpServletRequest request,
-                                                              @PathVariable(name = "name") String name
+    public ResponseEntity<ServerResponse> getServerByHostname(@PathVariable(name = "name") String name
     ) throws ManagedException {
 
         ServerResponse server = new ServerResponse(service.getServerByHostname(name));
