+++ src/main/java/com/br/kerberus/urd/entity/core/LogMetlhodReturn.java
@@ -1,4 +1,4 @@
-package com.br.kerberus.urd.entity;
+package com.br.kerberus.urd.entity.core;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,8 +7,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface LogExecutionTime {
+public @interface LogMetlhodReturn {
 
-    LogType LogType() default LogType.EXECUTION_TIME;
-
+    LogType LogType() default LogType.METHOD_RETURN;
 }
