+++ src/main/java/com/br/kerberus/urd/service/ApplicationService.java
@@ -1,6 +1,10 @@
 package com.br.kerberus.urd.service;
 
 import com.br.kerberus.urd.entity.*;
+import com.br.kerberus.urd.entity.core.LogException;
+import com.br.kerberus.urd.entity.core.LogExecutionTime;
+import com.br.kerberus.urd.entity.core.LogMethodCall;
+import com.br.kerberus.urd.entity.core.LogMetlhodReturn;
 import com.br.kerberus.urd.exception.ManagedException;
 import com.br.kerberus.urd.repository.ApplicationRepository;
 import com.br.kerberus.urd.resource.ApplicationRequest;
