+++ src/main/java/com/br/kerberus/urd/service/ServerService.java
@@ -1,6 +1,7 @@
 package com.br.kerberus.urd.service;
 
 import com.br.kerberus.urd.entity.*;
+import com.br.kerberus.urd.entity.core.*;
 import com.br.kerberus.urd.exception.ManagedException;
 import com.br.kerberus.urd.exception.NoManagedException;
 import com.br.kerberus.urd.repository.ServerRepository;
