+++ src/test/resources/application.yml
@@ -45,6 +45,7 @@ logging:
       urd: com.br.kerberus.urd, com.br.kerberus.urd.service, com.br.kerberus.urd.config, com.br.kerberus.urd.controller, com.br.kerberus.urd.core, com.br.kerberus.urd.repository, com.br.kerberus.urd.exception, com.br.kerberus.urd.core.log
       sql: org.hibernate, org.hibernate.sql, org.springframework.jdbc.core, org.flywaydb.core, org.flywaydb.core, org.hibernate.type
       web: org.springframework.core.codec, org.springframework.http, org.springframework.web
+      web-request: org.springframework.web.filter.CommonsRequestLoggingFilter
    path: "logs"
    pattern:
       console: "# ${hostname} | ${spring.application.name} | %d{dd-MM-yyyy HH:mm:ss.SSS} |  %highlight(%-5level) | %X(operationType)  | %logger{0} | %msg%n"
@@ -53,3 +54,4 @@ logging:
       urd: info
       web: error
       sql: error
+      web-request: debug
